refactor: auth 관련 로직 제거 및 리팩토링
