--- src/main/java/com/mfc/memberservice/common/config/ApplicationConfig.java
@@ -1,52 +1,0 @@
-package com.mfc.memberservice.common.config;
-
-import java.util.UUID;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-import com.mfc.memberservice.common.exception.BaseException;
-import com.mfc.memberservice.common.response.BaseResponseStatus;
-import com.mfc.memberservice.member.infrastructure.MemberRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-public class ApplicationConfig {
-	private final MemberRepository memberRepository;
-
-	@Bean
-	public UserDetailsService userDetailsService() {
-		return uuid -> memberRepository.findByUuid(uuid)
-				.orElseThrow(() -> new BaseException(BaseResponseStatus.MEMBER_NOT_FOUND));
-	}
-
-	@Bean
-	public AuthenticationProvider authenticationProvider() {
-		DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
-		authenticationProvider.setUserDetailsService(userDetailsService());
-		authenticationProvider.setPasswordEncoder(bCryptPasswordEncoder());
-
-		return authenticationProvider;
-	}
-
-	@Bean
-	public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
-		return configuration.getAuthenticationManager();
-	}
-
-	// 비밀번호 암호화
-	@Bean
-	public BCryptPasswordEncoder bCryptPasswordEncoder() {
-		return new BCryptPasswordEncoder();
-	}
-}
