--- src/main/java/com/mfc/memberservice/common/config/SecurityConfig.java
@@ -1,41 +1,0 @@
-package com.mfc.memberservice.common.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-import com.mfc.memberservice.common.jwt.JwtTokenProvider;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-	private final AuthenticationProvider authenticationProvider;
-
-	@Bean
-	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-		return http
-				.csrf(AbstractHttpConfigurer::disable)
-				.formLogin(AbstractHttpConfigurer::disable)
-				.authorizeHttpRequests((auth) -> auth
-						.requestMatchers("/**")
-						.permitAll()
-						.anyRequest()
-						.authenticated()
-				)
-				.authenticationProvider(authenticationProvider)
-				.sessionManagement((session) -> session
-						.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-				)
-				.build();
-	}
-}
