+++ src/main/java/com/mfc/memberservice/common/jwt/JwtTokenProvider.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import com.mfc.memberservice.common.config.RedisConfig;
-import com.mfc.memberservice.member.domain.Member;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
