+++ src/main/java/com/mfc/memberservice/member/application/MemberServiceImpl.java
@@ -2,27 +2,20 @@ package com.mfc.memberservice.member.application;
 
 import static com.mfc.memberservice.common.response.BaseResponseStatus.*;
 
-import java.util.Optional;
-
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.mfc.memberservice.common.exception.BaseException;
 import com.mfc.memberservice.common.jwt.JwtTokenProvider;
-import com.mfc.memberservice.member.domain.Member;
+import com.mfc.memberservice.member.domain.FavoriteStyle;
 import com.mfc.memberservice.member.domain.Partner;
 import com.mfc.memberservice.member.domain.User;
 import com.mfc.memberservice.member.dto.req.ModifyFavoriteStyleReqDto;
-import com.mfc.memberservice.member.dto.req.ModifyMemberReqDto;
 import com.mfc.memberservice.member.dto.resp.FavoriteStyleRespDto;
-import com.mfc.memberservice.member.dto.resp.ProfileRespDto;
 import com.mfc.memberservice.member.dto.resp.SignInRespDto;
-import com.mfc.memberservice.member.infrastructure.MemberRepository;
+import com.mfc.memberservice.member.infrastructure.FavoriteStyleRepository;
 import com.mfc.memberservice.member.infrastructure.PartnerRepository;
 import com.mfc.memberservice.member.infrastructure.UserRepository;
-import com.mfc.memberservice.member.domain.FavoriteStyle;
-import com.mfc.memberservice.member.infrastructure.FavoriteStyleRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -30,13 +23,19 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Transactional
 public class MemberServiceImpl implements MemberService {
-	private final MemberRepository memberRepository;
 	private final UserRepository userRepository;
 	private final PartnerRepository partnerRepository;
 	private final FavoriteStyleRepository favoriteStyleRepository;
 
-	private final PasswordEncoder encoder;
 	private final JwtTokenProvider tokenProvider;
+
+	@Transactional(readOnly = true)
+	@Override
+	public boolean verifyNickname(String nickname) {
+		return (userRepository.findByNickname(nickname).isEmpty()
+				&& partnerRepository.findByNickname(nickname).isEmpty());
+	}
+
 
 	@Override
 	public void modifyNickname(String uuid, String role, String nickname) {
@@ -55,26 +54,10 @@ public class MemberServiceImpl implements MemberService {
 	}
 
 	@Override
-	public void modifyPassword(String uuid, ModifyMemberReqDto dto) {
-		Member member = memberRepository.findByUuid(uuid)
-				.orElseThrow(() -> new BaseException(MEMBER_NOT_FOUND));
-
-		memberRepository.save(Member.builder()
-				.id(member.getId())
-				.email(member.getEmail())
-				.name(member.getName())
-				.phone(member.getPhone())
-				.password(encoder.encode(dto.getPassword()))
-				.status(member.getStatus())
-				.build()
-		);
-	}
-
-	@Override
 	public void modifyFavoriteStyle(String uuid, ModifyFavoriteStyleReqDto dto) {
 		favoriteStyleRepository.deleteByUuid(uuid);
 
-		dto.getFavoriteStyles().stream()
+		dto.getFavoriteStyles()
 				.forEach(i -> favoriteStyleRepository.save(
 						FavoriteStyle.builder()
 								.uuid(uuid)
@@ -89,25 +72,6 @@ public class MemberServiceImpl implements MemberService {
 						.stream()
 						.toList())
 				.build();
-	}
-
-	@Override
-	public void resign(String uuid) {
-		Member member = memberRepository.findByUuid(uuid)
-				.orElseThrow(() -> new BaseException(MEMBER_NOT_FOUND));
-
-		memberRepository.save(Member.builder()
-				.id(member.getId())
-				.email(member.getEmail())
-				.name(member.getName())
-				.phone(member.getPhone())
-				.password(member.getPassword())
-				.status((short)0)
-				.build());
-
-		userRepository.deleteByUuid(uuid);
-		partnerRepository.deleteByUuid(uuid);
-		favoriteStyleRepository.deleteByUuid(uuid);
 	}
 
 	@Override
@@ -144,13 +108,13 @@ public class MemberServiceImpl implements MemberService {
 				.id(user.getId())
 				.nickname(nickname)
 				.profileImage(user.getProfileImage())
-				.topSize(user.getTopSize().toString())
+				.topSize(user.getTopSize() == null ? null : user.getTopSize().toString())
 				.height(user.getHeight())
 				.weight(user.getWeight())
 				.imageAlt(user.getImageAlt())
 				.shoeSize(user.getShoeSize())
-				.bottomSize(user.getBottomSize().toString())
-				.bodyType(user.getBodyType().toString())
+				.bottomSize(user.getBottomSize() == null ? null : user.getBottomSize().toString())
+				.bodyType(user.getBodyType() == null ? null : user.getBodyType().toString())
 				.build()
 		);
 	}
