+++ src/main/java/com/mfc/memberservice/member/application/PartnerService.java
@@ -2,10 +2,8 @@ package com.mfc.memberservice.member.application;
 
 import com.mfc.memberservice.member.dto.req.CareerReqDto;
 import com.mfc.memberservice.member.dto.req.ModifyPartnerReqDto;
-import com.mfc.memberservice.member.dto.req.OptionReqDto;
 import com.mfc.memberservice.member.dto.req.UpdateSnsReqDto;
 import com.mfc.memberservice.member.dto.resp.CareerListRespDto;
-import com.mfc.memberservice.member.dto.resp.OptionListRespDto;
 import com.mfc.memberservice.member.dto.resp.PartnerAccountRespDto;
 import com.mfc.memberservice.member.dto.resp.PartnerPortfolioRespDto;
 import com.mfc.memberservice.member.dto.resp.ProfileRespDto;
@@ -19,10 +17,6 @@ public interface PartnerService {
 	void updateCareer(String uuid, Long careerId, CareerReqDto dto);
 	void deleteCareer(String uuid, Long careerId);
 	CareerListRespDto getCareerList(String partnerId);
-	void createOption(String uuid, OptionReqDto dto);
-	void updateOption(String partnerId, Long optionId, OptionReqDto dto);
-	void deleteOption(String partnerId, Long optionId);
-	OptionListRespDto getOptionList(String partnerId);
 	void updateDescription(String uuid, ModifyPartnerReqDto dto);
 	void updateAccount(String uuid, ModifyPartnerReqDto dto);
 	void updateAverageTime(String uuid, ModifyPartnerReqDto dto);
