--- src/main/java/com/mfc/memberservice/member/dto/req/ModifyMemberReqDto.java
@@ -1,8 +1,0 @@
-package com.mfc.memberservice.member.dto.req;
-
-import lombok.Getter;
-
-@Getter
-public class ModifyMemberReqDto {
-	private String password;
-}
