--- src/main/java/com/mfc/memberservice/member/dto/req/OptionReqDto.java
@@ -1,10 +1,0 @@
-package com.mfc.memberservice.member.dto.req;
-
-import lombok.Getter;
-
-@Getter
-public class OptionReqDto {
-	private String value;
-	private Integer price;
-	private String description;
-}
