--- src/main/java/com/mfc/memberservice/member/dto/req/SignInReqDto.java
@@ -1,9 +1,0 @@
-package com.mfc.memberservice.member.dto.req;
-
-import lombok.Getter;
-
-@Getter
-public class SignInReqDto {
-	private String email;
-	private String password;
-}
