--- src/main/java/com/mfc/memberservice/member/dto/req/SmsReqDto.java
@@ -1,9 +1,0 @@
-package com.mfc.memberservice.member.dto.req;
-
-import lombok.Getter;
-
-@Getter
-public class SmsReqDto {
-	private String phone;
-	private String code;
-}
