--- src/main/java/com/mfc/memberservice/member/vo/req/SignInReqVo.java
@@ -1,9 +1,0 @@
-package com.mfc.memberservice.member.vo.req;
-
-import lombok.Getter;
-
-@Getter
-public class SignInReqVo {
-	private String email;
-	private String password;
-}
