--- src/main/java/com/mfc/memberservice/member/vo/req/SmsReqVo.java
@@ -1,9 +1,0 @@
-package com.mfc.memberservice.member.vo.req;
-
-import lombok.Getter;
-
-@Getter
-public class SmsReqVo {
-	private String phone;
-	private String code;
-}
