--- src/main/java/com/mfc/memberservice/member/vo/resp/OptionListRespVo.java
@@ -1,10 +1,0 @@
-package com.mfc.memberservice.member.vo.resp;
-
-import java.util.List;
-
-import lombok.Getter;
-
-@Getter
-public class OptionListRespVo {
-	private List<OptionVO> options;
-}
