--- src/main/java/com/mfc/memberservice/member/vo/resp/OptionVO.java
@@ -1,11 +1,0 @@
-package com.mfc.memberservice.member.vo.resp;
-
-import lombok.Getter;
-
-@Getter
-public class OptionVO {
-	private Long optionId;
-	private String value;
-	private Integer price;
-	private String description;
-}
